- Moved the helper function to correct file
